requirements.txt
flask
flask_restplus
Werkzeug==0.16.0
pymysql
sqlacodegen
sqlalchemy
flask_restful
flask-cors
python-dotenv
flask_sqlalchemy
pylint-flask-sqlalchemy
pylint-flask